Create PLEASE_LOOK_AT_THIS_DATA.txt
ATTENTION !!! 
I uploaded the file for averages by accident. The intention was to upload the tidy data file. I have provided the top 10 rows of the tidy data and also a link for the tidy data zipped file.

I request you to kindly consider this when you are evaluating/grading my submission.

Thanks!